+++ geothermsim/borefield/borefield.py
@@ -155,7 +155,7 @@ class Borefield:
         return h_to_point
 
     @classmethod
-    def from_dimensions(cls, L: ArrayLike, D: ArrayLike, r_b: ArrayLike, x: ArrayLike, y: ArrayLike, basis: Basis, n_segments: int, tilt: ArrayLike = 0., orientation: ArrayLike = 0., segment_ratios: ArrayLike | None = None) -> Self:
+    def from_dimensions(cls, L: ArrayLike, D: ArrayLike, r_b: ArrayLike, x: ArrayLike, y: ArrayLike, basis: Basis, n_segments: int, tilt: ArrayLike = 0., orientation: ArrayLike = 0., segment_ratios: ArrayLike | None = None, order: int = 101, order_to_self: int = 21) -> Self:
         """Field of straight boreholes from their dimensions.
 
         Parameters
@@ -185,6 +185,15 @@ class Borefield:
             (i.e. ``sum(segment_ratios) = 1``). If `segment_ratios` is
             ``None``, segments of equal size are considered (i.e.
             ``segment_ratios[v] = 1 / n_segments``).
+        order : int, default: 101
+            Order of the Gauss-Legendre quadrature to evaluate thermal
+            response factors to points outside the borehole, and to evaluate
+            coefficient matrices for fluid and heat extraction rate profiles.
+        order_to_self : int, default: 21
+            Order of the tanh-sinh quadrature to evaluate thermal
+            response factors to nodes on the borehole. Corresponds to the
+            number of quadrature points along each subinterval delimited
+            by nodes and edges of the segments.
 
         Returns
         -------
@@ -210,11 +219,11 @@ class Borefield:
         boreholes = []
         for j in range(n_boreholes):
             path = Path.Line(L[j], D[j], x[j], y[j], tilt[j], orientation[j])
-            boreholes.append(Borehole(r_b[j], path, basis, n_segments, segment_ratios=segment_ratios))
+            boreholes.append(Borehole(r_b[j], path, basis, n_segments, segment_ratios=segment_ratios, order=order, order_to_self=order_to_self))
         return cls(boreholes)
 
     @classmethod
-    def rectangle_field(cls, N_1: int, N_2: int, B_1: float, B_2: float, L: float, D: float, r_b: float, basis: Basis, n_segments: int, segment_ratios: ArrayLike | None = None, order: int | None = None) -> Self:
+    def rectangle_field(cls, N_1: int, N_2: int, B_1: float, B_2: float, L: float, D: float, r_b: float, basis: Basis, n_segments: int, segment_ratios: ArrayLike | None = None, order: int = 101, order_to_self: int = 21) -> Self:
         """Field of vertical boreholes in a rectangular configuration.
 
         Parameters
@@ -238,6 +247,15 @@ class Borefield:
             (i.e. ``sum(segment_ratios) = 1``). If `segment_ratios` is
             ``None``, segments of equal size are considered (i.e.
             ``segment_ratios[v] = 1 / n_segments``).
+        order : int, default: 101
+            Order of the Gauss-Legendre quadrature to evaluate thermal
+            response factors to points outside the borehole, and to evaluate
+            coefficient matrices for fluid and heat extraction rate profiles.
+        order_to_self : int, default: 21
+            Order of the tanh-sinh quadrature to evaluate thermal
+            response factors to nodes on the borehole. Corresponds to the
+            number of quadrature points along each subinterval delimited
+            by nodes and edges of the segments.
 
         Returns
         -------
@@ -248,5 +266,5 @@ class Borefield:
         # Borehole positions and orientation
         x = jnp.tile(jnp.arange(N_1), N_2) * B_1
         y = jnp.repeat(jnp.arange(N_2), N_1) * B_2
-        return cls.from_dimensions(L, D, r_b, x, y, basis, n_segments, segment_ratios=segment_ratios)
+        return cls.from_dimensions(L, D, r_b, x, y, basis, n_segments, segment_ratios=segment_ratios, order=order, order_to_self=order_to_self)